fd-tracker: add epoll/poll management wrappers to fd-tracker
Add wrappers which create an lttng_poll_event object, tracking
any file descriptor created in the process of initializing
the object.

When the build is configured to use the epoll interface, the
underlying epoll fd is tracked by the fd-tracker.

If the build is configured to use the poll interace, the fd tracker is
not involved in the process and the utility will simply defer the call
to lttng_poll_create() directly.

Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>