fd-tracker: add the unlink operation to fs handles
The unlink method allows user to unlink the file refered-to by
an fs_handle. As indicated in the comments, the unlink operation
leverages the lttng_inode to ensure the actual unlink system call
is only performed when the last reference to an lttng_inode is
released by an fs_handle.

Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>