fd-tracker: add the lttng-inode interface
The lttng-inode interface is internal to the fd-tracker. It allows
unlink operations on an fs_handle to determine if another fs_handle
refers to the same file.

When this is deemed to be the case, the unlink can be delayed by
renaming the file to a temporary name and, finally, unlinking it
when all references have been released.

Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>