Format lists in src/common/Makefile.am
I'm currently dealing with src/common/Makefile.am, and would find it
more pleasant to modify if the lists were formatted consistently and
sorted (except DIST_SUBDIRS, for which the order is important).

I have taken the liberty to put the corresponding .c and .h on the same
line, since they will always be next to each other.  It reduces a bit
the number of lines.  If you don't like that, I would also be fine with
a completely flat list.

Change-Id: I8a65fe5e0fea37230f1376c51c7d4a547f651c7e
Signed-off-by: Simon Marchi <simon.marchi@efficios.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>