Silence bogus warning when building with old GCC versions
gcc version 7.4.0 (and possibly others) warn that `new_handle`
may be used uninitialized. It is not clear what code path would
trigger this uninitialized use.

Moreover, the warning is not present when building with recent
versions of gcc and clang, nor is it flagged by Coverity. This points
to the warning being bogus.

Still, setting `new_handle` to NULL silences the warning on this
gcc version.

Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>