Fix: unchecked return value of lttng_directory_handle_create()
Although unlikely in this case (as the path was just created), the
creation of a directory handle can fail and this should always be
accounted-for.

1415129 Dereference null return value

If the function actually returns a null value, a null pointer
dereference will occur.

In lttng_unlinked_file_pool_add_inode: Return value of function which
returns null is dereferenced without checking (CWE-476)

Reported-by: Coverity Scan
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>
Change-Id: I914802c25b569320e08265aae251e062bccbd50e