Clean-up: trace-chunk: remove unreachable code
`path` can never be NULL in this code path; it is unnecessary to check
it.

There may be a null pointer dereference, or else the comparison
against null is unnecessary.

In lttng_trace_chunk_rename_path_no_lock: All paths that lead to this
null pointer comparison already dereference the pointer
earlier (CWE-476).

Reported-by: Coverity Scan
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>
Change-Id: I4d850b0b8686300268d85f577ae5f14a96d66348