tests: append to AM_CFLAGS instead of overriding it
The Makefiles modified by this patch currently override the AM_CFLAGS
value, which means that anything put in AM_CFLAGS by configure (for
example, the warning flags) is lost.

I believe the intention is to add some flags to CFLAGS, so modify them
such that they append instead of override.

notification/Makefile.am overrides AM_LDFLAGS with nothing.  It feels
like it was not the intention to actually clear that variable, but
rather that it is just the by-product of a copy paste.  If it was really
the intention to clear the value of AM_LDFLAGS, there would have been a
comment to explain it, right?

Signed-off-by: Simon Marchi <simon.marchi@efficios.com>
Change-Id: I4ef926d9135b16200e5f17d09461506a5e955068
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>