configure: add --enable-Werror
If one wants to build with -Werror, it's not possible to simply
configure with -Werror in the CFLAGS.  This makes a bunch of configure
checks fail, which would have otherwise passed.

This patch adds an --enable-Werror option to configure, which has the
effect of adding -Werror to AM_CFLAGS.  It therefore ends up in the
CFLAGS used to build the project, but it doesn't interfere with the
configure checks.

Signed-off-by: Simon Marchi <simon.marchi@efficios.com>
Change-Id: I18c33125c717305aac8f1d8a19fee7e065d70c31
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>