Fix: Tests: `gen-ust-events` doesn't error out on invalid option
Issue
=====
When running `gen-ust-events` with an invalid option:
  ./gen-ust-events -h

The `getopt_long()` function prints the following error:
  ./gen-ust-events: invalid option -- 'h'

which is very kind of it.

The problem is that the process keep running and go on to generate
events. It should exit right away.

Solution
========
Remove the `break` statement and so that we execute the `goto end`
right away.

Apply the same changes to `gen-ust-nevents`.

Signed-off-by: Francis Deslauriers <francis.deslauriers@efficios.com>
Change-Id: Ib1c96f4c9ed8f98395bf842215f858a69db2bbf0
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>