Fix: tests: add `void` parameter to functions that take no parameters
In C, empty parenthesis declare a function without a prorotype (without
specifying its parameters).  This is not the same as having a `void`
parameter, which declares a function which has no parameters.

It's safer to use the later, otherwise it makes it possible to
erroneously call the function with some arguments.

Change this `test_function` to add `void`.  It fixes diagnostics like:

  CC       userspace-probe-elf-binary.o
/home/simark/src/lttng-tools/tests/utils/testapp/userspace-probe-elf-binary/userspace-probe-elf-binary.c:12:33: error: no previous prototype for ‘test_function’ [-Werror=missing-prototypes]
   12 | void __attribute__ ((noinline)) test_function()
      |

Change-Id: Iceb7636e44d45f51889667ec76f2c04c032b5df8
Signed-off-by: Simon Marchi <simon.marchi@efficios.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>