Fix: python: suppress -Wmissing-prototypes warning with SWIG 3.0.10
SWIG 3.0.10 is used on SLES 12.  It produces this warning:

      CC       lttng_wrap.lo
    lttng_wrap.c:3411:1: error: no previous prototype for ‘SWIG_strnlen’ [-Werror=missing-prototypes]
     3411 | SWIG_strnlen(const char* s, size_t maxlen)
          | ^~~~~~

SWIG_strnlen is defined like this:

    size_t
    SWIG_strnlen(const char* s, size_t maxlen)
    {
        ...
    }

Since the function is not static and has no previous declaration, the
diagnostic is emitted.  We can see that they have fixed it in SWIG
3.0.12, where the same function is defined as:

    SWIGINTERN size_t
    SWIG_strnlen(const char* s, size_t maxlen)
    {
        ...
    }

SWIGINTERN is defined as static.

We can workaround the warning by adding our own declaration for
SWIG_strnlen in lttng.i(.in).

Tested by build with SWIG 3.0.10, 3.0.11 and 3.0.12.

Change-Id: I7d9d93cf5fc04f2044a47ad384e16d726e27f72f
Signed-off-by: Simon Marchi <simon.marchi@efficios.com>
Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>